Array Housekeeping and channel ID
Added removing of post in the array if it gets too large.
Started possible integration of using channel ID to post in the proper place.
TODO: Instead of channel ID's, would be better to use already created webhooks on the server to connect to and post as.
NOTE: Both features is untested...