Animated emojis will be sent by the bot
Bot can detect valid emojis in a user's message and will resend the message with a webhook with working emojis.
These emojis have to be animated and exist in the server, otherwise, the bot will not do anything with it.
Bot uses regular expression to recognize which words are the emojis